Backend - Ajustes

+++ database/migrations/tenant/2025_05_28_150500_create_cards_installments_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
                 $table->string('status')->nullable(false); // Ex: 'open', 'closed', 'paid', 'overdue'
                 $table->decimal('amount', 15)->nullable(false);
                 $table->integer('installment')->nullable(false);
-                $table->integer('installment_amount')->nullable(false);
+                $table->decimal('installment_amount', 15)->nullable(false);
                 $table->string('date')->nullable(false);
                 $table->boolean('deleted')->default(false);
